Move pure functions to utils.js
